Reworking this cookbook to get it working
Replacing the nexus.erb a copy of the Nexus' bin/nexus file with additional configuration to run correctly.
Minor changes to the recipe to remove some unneeded things.
Change to the recipe to copy nginx config files and get it working.
Adding a self-signed cert that should be replaced with a real cert in production.
Updated README
Added .gitignore